services/user-service/requirements.txt
# Flask 核心框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# 数据库
PyMySQL==1.1.0
cryptography==41.0.4

# Redis
redis==4.6.0

# JWT 认证
PyJWT==2.8.0

# 密码加密
Werkzeug==2.3.7

# Nacos 服务发现
nacos-sdk-python==0.1.5

# HTTP 客户端
requests==2.31.0

# 配置管理
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 数据验证
marshmallow==3.20.1

# 时间处理
python-dateutil==2.8.2

# 开发工具
flake8==6.0.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# WSGI 服务器
gunicorn==21.2.0